[fix]: search page에서 재사용을 위해 path 분기처리: 진행중
